+++ pkg/nats/common.go
@@ -41,6 +41,10 @@ import (
 )
 
 type errorFormatterService interface {
+	// ErrorNoWrap function for pseudo-wrapp error, must be used in case of linter warnings...
+	ErrorNoWrap(err error) error
+	// ErrNoWrap same with ErrorNoWrap function, just alias for ErrorNoWrap, just short function name...
+	ErrNoWrap(err error) error
 	ErrorOnly(err error, details ...string) error
 	Errorf(err error, format string, args ...interface{}) error
 	NewError(details ...string) error